requirements.txt
numpy
spidev
RPi.GPIO
Pillow
inky
Flask
bluepy
pybluez
beacontools[scan]
